+++ src/utils/url/index.ts
@@ -1,0 +1,1 @@
+export const getBungieApiUrl = (url: string) => `https://www.bungie.net${url}`;
